Wrote first post for ongoing journal for the MDM method
Wrote the first post for the MDM journal I am going to be writing th throughout the duration of this project. Interestingly, I wrote this on Google Keep before confirming/figuring out the workflow I wanted to use with GitHub. Might still have Keep notes I write when away from my desktop (and maybe a folder for audio recordings and images or something like that).